strapi xstate sessions

+++ src/pages/logout.js
@@ -1,31 +1,28 @@
-import React from "react"
-import { Row, Col, Container } from "react-bootstrap"
-import Layout from "src/components/layout"
+import React, { useContext } from "react"
+import { store } from "src/store"
 import SEO from "src/components/seo"
-import useLocalStorage from "src/hooks/useLocalStorage"
-import { useLocation, useNavigate } from "@reach/router"
+import { navigate } from "gatsby"
+import useIsClient from "src/hooks/use-is-client"
 
-const Logout = ({ data }) => {
-  const navigate = useNavigate()
-  const [session, setSession] = useLocalStorage()
-  setTimeout(() => {
-    setSession(false)
+const Logout = () => {
+  const { isClient, key } = useIsClient()
+  const sessionMachine = useContext(store)
+  const { state, send } = sessionMachine
+  if (isClient) {
     window.localStorage.clear()
-    navigate("/", { replace: true })
-  }, 500)
-  let text = "Logging you out…"
+    send("LOGOUT")
+    setTimeout(() => {
+      navigate("/")
+    }, 500)
+  }
   return (
-    <Layout pageInfo={{ pageName: "login" }}>
-      <SEO title="Home" keywords={["Linceo", "Young"]} />
-      <Container className="">
-        <Row>
-          <Col>
-            {text}
-            {/* <pre>{JSON.stringify(data, null,2)}</pre> */}
-          </Col>
-        </Row>
-      </Container>
-    </Layout>
+    <>
+      <SEO title="Logout" />
+      <div className="p-5">
+        <h1>Goodbye!</h1>
+        <p>See you next time.</p>
+      </div>
+    </>
   )
 }
